move fethcing into module for seperation of concern
